Simplify adding icons to the desktop
Use xfdesktop_grid_get_next_free_position when placing icons on
the desktop rather than having it spead over multiple functions.

+++ src/xfdesktop-file-icon-manager.h
@@ -59,8 +59,8 @@ gboolean xfdesktop_file_icon_manager_get_cached_icon_position(
                                                     XfdesktopFileIconManager *fmanager,
                                                     const gchar *name,
                                                     const gchar *identifier,
-                                                    gint16 *row,
-                                                    gint16 *col);
+                                                    gint *row,
+                                                    gint *col);
 
 G_END_DECLS
